+++ modules/Sources/Features/Root/RootStore.swift
@@ -130,9 +130,6 @@ public struct Root {
         public var autoUpdateSwapCandidates: IdentifiedArrayOf<TransactionState> = []
         @Shared(.inMemory(.swapAssets)) public var swapAssets: IdentifiedArrayOf<SwapAsset> = []
 
-        // Shared Flags
-        @Shared(.inMemory(.sharedFlags)) public var sharedFlags: SharedFlags? = nil
-
         // Path
         public var addKeystoneHWWalletCoordFlowState = AddKeystoneHWWalletCoordFlow.State.initial
         public var currencyConversionSetupState = CurrencyConversionSetup.State.initial
@@ -277,10 +274,6 @@ public struct Root {
         case checkFundsFoundSomething
         case checkFundsNothingFound
         case checkFundsTorRequired
-        
-        // Shared flags
-        case closeInsufficientFundsTapped
-        case updateSharedFlags(SharedFlags?)
     }
 
     @Dependency(\.addressBook) var addressBook
@@ -458,14 +451,6 @@ public struct Root {
 
             case .onboarding(.newWalletSuccessfulyCreated):
                 return .send(.initialization(.initializeSDK(.newWallet)))
-
-            case .closeInsufficientFundsTapped:
-                state.$sharedFlags.withLock { $0 = nil }
-                return .none
-                
-            case .updateSharedFlags(let newValue):
-                state.$sharedFlags.withLock { $0 = newValue }
-                return .none
 
             default: return .none
             }