+++ modules/Sources/Features/Settings/SettingsStore.swift
@@ -85,6 +85,8 @@ public struct Settings {
         case advancedSettingsTapped
         case binding(BindingAction<Settings.State>)
         case checkFundsForAddress(String)
+        case currencyConversionGranted
+        case currencyConversionTapped
         case enableEnhanceTransactionMode
         case enableRecoverFundsMode
         case fetchDataForTxid(String)
@@ -131,7 +133,17 @@ public struct Settings {
                         await send(.addressBookTapped)
                     }
                 }
+
+            case .currencyConversionTapped:
+                return .run { send in
+                    if await localAuthentication.authenticate() {
+                        await send(.currencyConversionGranted)
+                    }
+                }
                 
+            case .currencyConversionGranted:
+                return .none
+
             case .addressBookTapped:
                 return .none